Full working services to start/delte get status and get results job

+++ java/src/test/java/io/swagger/api/JobsApiControllerIntegrationTest.java
@@ -27,7 +27,7 @@ public class JobsApiControllerIntegrationTest {
 
     @Test
     public void jobsJobIdDeleteTest() throws Exception {
-        Integer jobId = 56;
+        Long jobId = 56L;
         ResponseEntity<Void> responseEntity = api.jobsJobIdDelete(jobId);
         assertEquals(HttpStatus.NOT_IMPLEMENTED, responseEntity.getStatusCode());
     }
